fix: improve logic

+++ app/src/main/java/com/acadify/model/data/KomponenNilai.kt
@@ -5,11 +5,11 @@ data class KomponenNilai(
     val nilaiKuis: Float = 0f,
     val nilaiUTS: Float = 0f,
     val nilaiUAS: Float = 0f,
-    val persentaseTugas: Int = 25,
-    val persentaseKuis: Int = 25,
-    val persentaseUTS: Int = 25,
-    val persentaseUAS: Int = 25,
+    val persentaseTugas: Float = 25f,
+    val persentaseKuis: Float = 25f,
+    val persentaseUTS: Float = 25f,
+    val persentaseUAS: Float = 25f,
 ) {
     // No-argument constructor for Firestore
-    constructor() : this(0f, 0f, 0f, 0f, 25, 25, 25, 25)
+    constructor() : this(0f, 0f, 0f, 0f, 25f, 25f, 25f, 25f)
 }
